+++ libs/package-template/src/example.module.spec.ts
@@ -1,5 +1,6 @@
-import { ExampleModule } from "@lib"
 import { Test } from "@nestjs/testing"
+
+import { ExampleModule } from "@lib"
 
 describe("ExampleModule", () => {
   it("should compile the module", async () => {